Removed pointless casts.  Patch by Chris Frey <cdfrey@netdirect.ca>
git-svn-id: svn://svn.gna.org/svn/mysqlpp/trunk@410 5946f024-35f4-0310-b1b3-9a3be3380cfb

+++ examples/cgi_image.cpp
@@ -23,7 +23,7 @@ main(int argc, char *argv[])
 	Connection con(use_exceptions);
 	try {
 		con.real_connect(MY_DATABASE, MY_HOST, MY_USER, MY_PASSWORD, 3306,
-						 (int) 0, 60, NULL);
+						 0, 60, NULL);
 		Query query = con.query();
 		query << "SELECT " << MY_FIELD << " FROM " << MY_TABLE << " WHERE "
 			<< MY_KEY << " = " << argv[1];
